requirements.txt
asgiref==3.5.2
cfgv==3.3.1
distlib==0.3.5
Django==4.0.6
django-environ==0.9.0
djangorestframework==3.13.1
filelock==3.7.1
gunicorn==20.1.0
identify==2.5.2
nodeenv==1.7.0
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2==2.9.3
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.15.1
